Inject admin helper methods via ad-hoc decorator
As for the frontend helper methods in the previous commit, using
`Spree::Admin::PaymentsController.helper` in
`config/initializers/braintree.rb` does not work: at `checkout/payement`
step an exception is raised.

```
ActionView::Template::Error (undefined method `braintree_transaction_link`)
```

This commit firstly move the helper module in the appropriate
namespace (i.e. `SolidusPaypalBraintree`), then inject into the
controller via a custom decorator. This way the exception is no longer raised.
