+++ app/controllers/solidus_paypal_braintree/admin_payments_controller_decorator.rb
@@ -1,0 +1,10 @@
+module SolidusPaypalBraintree
+  module AdminPaymentsControllerDecorator
+
+    def self.prepended(base)
+      base.helper ::SolidusPaypalBraintree::BraintreeAdminHelper
+    end
+
+    ::Spree::Admin::PaymentsController.prepend(self)
+  end
+end
