--- config/initializers/braintree.rb
@@ -1,3 +1,0 @@
-if SolidusSupport.backend_available?
-  Spree::Admin::PaymentsController.helper :braintree_admin
-end
